move information function
