+++ index.html
@@ -236,7 +236,7 @@
     <!-- navigation -->
     <script src="navigasi.js"></script>
     <script src="node_modules/clipboard/dist/clipboard.min.js"></script>
-    <script type="text/javascript">
+    <script>
         new ClipboardJS('.btn');
     </script>
 </body>
