+++ navigasi.js
@@ -279,3 +279,14 @@ function text_check_page(){
 	// Append to the `head` element
 	document.head.appendChild(script);
 }
+
+const {shell} = require('electron');
+function openDev(){
+    var url = "https://github.com/veler/DevToys";
+    shell.openExternal(url)
+}
+
+function openRepo(){
+    var url = "https://github.com/rzak23/FormDev";
+    shell.openExternal(url)
+}
